configure docker file

+++ src/main/java/com/ananda/AnandaJava/dao/UserDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ananda.AnandaJava.model.Users;
 
-public interface UserDao extends JpaRepository<Users, Long> {
+public interface UserDao extends JpaRepository<Users, Integer> {
 	
 	List<Users> findByFirstName(String name);
 	List<Users> findByEmailId(String reqBody);
