+++ src/main/java/com/ananda/AnandaJava/dao/UserRolesDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ananda.AnandaJava.model.UserRoles;
 
-public interface UserRolesDao extends JpaRepository<UserRoles, Long> {
+public interface UserRolesDao extends JpaRepository<UserRoles, Integer> {
 
 }
